Finished rough draft in network-culture-is-inevitable.md
I can already think of several ways to improve this writing and expand upon it, but frankly I'm struggling to focus on this right now, and I feel like it's cohesive enough to at least begin taking feedback on. I'll probably edit it before sharing it with the club meeting on Tuesday 2/11/2025.